@@ -1,6 +1,6 @@
 [tool.ruff]
 # The minimum Python version that should be supported
-target-version = "py39"
+target-version = "py311"
 
 line-length = 100
 
@@ -50,13 +50,13 @@ backend-path = ["."]
 name = "cvxpy"
 description = "A domain-specific language for modeling convex optimization problems in Python."
 dependencies = [
-    "osqp >= 0.6.2",
+    "osqp >= 1.0.0",
     "clarabel >= 0.5.0",
     "scs >= 3.2.4.post1",
-    "numpy >= 1.21.6",
-    "scipy >= 1.11.0",
+    "numpy >= 2.0.0",
+    "scipy >= 1.13.0",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 urls = {Homepage = "https://github.com/cvxpy/cvxpy"}
 license = {text = "Apache License, Version 2.0"}
 authors = [{name = "Steven Diamond", email = "stevend2@stanford.edu"},
@@ -69,26 +69,30 @@ dynamic = ["version"]
 # Solver names as in cvxpy.settings = pip-installable distribution providing it
 CBC = ["cylp>=0.91.5"]
 CLARABEL = []
+# CUOPT dependency group is disabled since it was breaking uv sync.
+# CUOPT = ["cuopt-cu12>=25.5", "nvidia-cuda-runtime-cu12>=12.8,<13.0"]
 CVXOPT = ["cvxopt"]
 DIFFCP = ["diffcp"]
 ECOS = ["ecos"]
 ECOS_BB = ["ecos"]
-GLOP = ["ortools>=9.7,<9.12"]
+GLOP = ["ortools>=9.7,<9.15"]
 GLPK = ["cvxopt"]
 GLPK_MI = ["cvxopt"]
 GUROBI = ["gurobipy"]
 HIGHS = ["highspy"]
 MOSEK = ["Mosek"]
 OSQP = []
-PDLP = ["ortools>=9.7,<9.12"]
+PDLP = ["ortools>=9.7,<9.15"]
 PIQP = ["piqp"]
 PROXQP = ["proxsuite"]
 QOCO = ["qoco"]
+QPALM = ["qpalm~=1.2.5"]
 SCIP = ["PySCIPOpt"]
 SCIPY = ["scipy"]
 SCS = []
-XPRESS = ["xpress"]
+XPRESS = ["xpress>=9.5"]
 DAQP = ["daqp"]
+KNITRO = ["knitro"]
 testing = ["pytest", "hypothesis"]
 doc = ["sphinx",
        "sphinxcontrib.jquery",
