@@ -70,8 +70,12 @@ class norm1(AxisAtom):
         return [self.axis]
 
     def name(self) -> str:
-        return "%s(%s)" % (self.__class__.__name__,
-                           self.args[0].name())
+        return f"{type(self).__name__}({self.args[0].name()})"
+
+    def format_labeled(self) -> str:
+        if self._label is not None:
+            return self._label
+        return f"{type(self).__name__}({self.args[0].format_labeled()})"
 
     def _domain(self) -> List[Constraint]:
         """Returns constraints describing the domain of the node.
