@@ -61,7 +61,12 @@ class one_minus_pos(Atom):
         return sp.csc_array(-1.0 * self._ones)
 
     def name(self) -> str:
-        return "%s(%s)" % (self.__class__.__name__, self.args[0])
+        return f"{self.__class__.__name__}({self.args[0]})"
+
+    def format_labeled(self) -> str:
+        if self._label is not None:
+            return self._label
+        return f"{self.__class__.__name__}({self.args[0].format_labeled()})"
 
     def shape_from_args(self) -> Tuple[int, ...]:
         """Returns the (row, col) shape of the expression.
