@@ -53,11 +53,36 @@ class BinaryOperator(AffAtom):
 
     def name(self):
         pretty_args = []
-        for a in self.args:
+        for i, a in enumerate(self.args):
+            # Always parenthesize AddExpression and DivExpression
             if isinstance(a, (AddExpression, DivExpression)):
+                pretty_args.append('(' + a.name() + ')')
+            # For division, also parenthesize multiplication on the right
+            elif isinstance(self, DivExpression) and i == 1 and \
+                    isinstance(a, (MulExpression, multiply)):
                 pretty_args.append('(' + a.name() + ')')
             else:
                 pretty_args.append(a.name())
+        return pretty_args[0] + ' ' + self.OP_NAME + ' ' + pretty_args[1]
+    
+    def format_labeled(self):
+        """Format binary operation with labels where available."""
+        # Check for own label first
+        if self._label is not None:
+            return self._label
+        
+        # Build from sub-expressions using their labels
+        pretty_args = []
+        for i, a in enumerate(self.args):
+            # Always parenthesize AddExpression and DivExpression
+            if isinstance(a, (AddExpression, DivExpression)):
+                pretty_args.append('(' + a.format_labeled() + ')')
+            # For division, also parenthesize multiplication on the right
+            elif isinstance(self, DivExpression) and i == 1 and \
+                    isinstance(a, (MulExpression, multiply)):
+                pretty_args.append('(' + a.format_labeled() + ')')
+            else:
+                pretty_args.append(a.format_labeled())
         return pretty_args[0] + ' ' + self.OP_NAME + ' ' + pretty_args[1]
 
     def numeric(self, values):
