@@ -20,6 +20,7 @@ from typing import Optional, Tuple
 
 import numpy as np
 import scipy.sparse as sp
+from numpy.lib.array_utils import normalize_axis_index
 
 from cvxpy.atoms.atom import Atom
 
@@ -80,12 +81,9 @@ def partial_transpose(expr, dims: Tuple[int, ...], axis: Optional[int] = 0):
     expr = Atom.cast_to_const(expr)
     if expr.ndim < 2 or expr.shape[0] != expr.shape[1]:
         raise ValueError("partial_transpose only supports 2-d square arrays.")
-    if axis < 0 or axis >= len(dims):
-        raise ValueError(
-            f"Invalid axis argument, should be between 0 and {len(dims)}, got {axis}."
-        )
     if expr.shape[0] != np.prod(dims):
         raise ValueError("Dimension of system doesn't correspond to dimension of subsystems.")
+    axis = normalize_axis_index(axis, len(dims))
     return sum([
         _term(expr, i, j, dims, axis) for i in range(dims[axis]) for j in range(dims[axis])
     ])
