@@ -20,10 +20,26 @@ import numpy as np
 import cvxpy.lin_ops.lin_op as lo
 import cvxpy.lin_ops.lin_utils as lu
 from cvxpy.atoms.affine.affine_atom import AffAtom
+from cvxpy.atoms.affine.binary_operators import MulExpression
 from cvxpy.constraints.constraint import Constraint
 
 
-class trace(AffAtom):
+def trace(expr):
+    """
+    TLDR: Use alternate formulation for trace(A@B) for more efficient computation.
+    trace(A@B) normally is O(n^3) because of the A@B operation. 
+    However, trace(A@B) only requires diagonal entries of A@B, which can be
+    computed by taking the sum of element-wise product of A.T * B in O(n^2) time.
+    In fact, vdot does this operation more robustly, using conj(A) instead of transpose. 
+    """
+    if isinstance(expr, MulExpression): 
+        from cvxpy.atoms.affine.binary_operators import vdot
+        return vdot(expr.args[0], expr.args[1])
+    else:
+        return Trace(expr)
+
+
+class Trace(AffAtom):
     """The sum of the diagonal entries of a matrix.
 
     Parameters
@@ -33,7 +49,7 @@ class trace(AffAtom):
     """
 
     def __init__(self, expr) -> None:
-        super(trace, self).__init__(expr)
+        super(Trace, self).__init__(expr)
 
     def sign_from_args(self) -> Tuple[bool, bool]:
         """Trace is nonneg (nonpos) if its argument is elementwise nonneg
