@@ -113,6 +113,7 @@ class Leaf(expression.Expression):
                 raise ValueError("Invalid dimensions %s." % (shape,))
         shape = tuple(shape)
         self._shape = shape
+        super(Leaf, self).__init__()
 
         if (PSD or NSD or symmetric or diag or hermitian) and (len(shape) != 2
                                                                or shape[0] != shape[1]):
@@ -403,14 +404,14 @@ class Leaf(expression.Expression):
             return val.astype(complex)
         elif self.attributes['boolean']:
             if hasattr(self, "boolean_idx"):
-                new_val = val.copy()
-                new_val[self.boolean_idx] = np.round(np.clip(val[self.boolean_idx], 0., 1.))
-                return new_val
+                new_val = np.atleast_1d(val.astype(np.float64, copy=True))
+                new_val[self.boolean_idx] = np.round(np.clip(new_val[self.boolean_idx], 0., 1.))
+                return new_val.reshape(val.shape) if val.ndim == 0 else new_val
         elif self.attributes['integer']:
             if hasattr(self, "integer_idx"):
-                new_val = val.copy()
-                new_val[self.integer_idx] = np.round(val[self.integer_idx])
-                return new_val
+                new_val = np.atleast_1d(val.astype(np.float64, copy=True))
+                new_val[self.integer_idx] = np.round(new_val[self.integer_idx])
+                return new_val.reshape(val.shape) if val.ndim == 0 else new_val
         elif self.attributes['diag']:
             if intf.is_sparse(val):
                 val = val.diagonal()
