@@ -17,7 +17,7 @@ limitations under the License.
 from cvxpy.atoms import (MatrixFrac, Pnorm, QuadForm, abs, bmat, conj, conv,
                          cumsum, imag, kron, lambda_max, lambda_sum_largest,
                          log_det, norm1, norm_inf, quad_over_lin, real,
-                         reshape, sigma_max, trace, upper_tri,
+                         reshape, sigma_max, Trace, upper_tri,
                          von_neumann_entr, quantum_rel_entr)
 from cvxpy.atoms.affine.add_expr import AddExpression
 from cvxpy.atoms.affine.binary_operators import (DivExpression, MulExpression,
@@ -74,7 +74,7 @@ CANON_METHODS = {
     broadcast_to: separable_canon,
     reshape: separable_canon,
     Sum: separable_canon,
-    trace: trace_canon,
+    Trace: trace_canon,
     transpose: separable_canon,
     NegExpression: separable_canon,
     upper_tri: separable_canon,
