@@ -180,17 +180,17 @@ class XPRESS(ConicSolver):
 
         self.prob_.loadproblem(probname="CVX_xpress_conic",
                                # constraint types
-                               qrtypes=['E'] * nrowsEQ + ['L'] * nrowsLEQ,
+                               rowtype=['E'] * nrowsEQ + ['L'] * nrowsLEQ,
                                rhs=b,                               # rhs
-                               range=None,                          # range
-                               obj=c,                               # obj coeff
-                               mstart=mstart,                       # mstart
-                               mnel=None,                           # mnel (unused)
+                               rng=None,                            # range
+                               objcoef=c,                           # obj coeff
+                               start=mstart,                        # mstart
+                               collen=None,                         # mnel (unused)
                                # linear coefficients
-                               mrwind=A.indices[A.data != 0],       # row indices
-                               dmatval=A.data[A.data != 0],         # coefficients
-                               dlb=[-xp.infinity] * len(c),         # lower bound
-                               dub=[xp.infinity] * len(c),          # upper bound
+                               rowind=A.indices[A.data != 0],       # row indices
+                               rowcoef=A.data[A.data != 0],         # coefficients
+                               lb=[-xp.infinity] * len(c),          # lower bound
+                               ub=[xp.infinity] * len(c),           # upper bound
                                colnames=varnames,                   # column names
                                rownames=linRownames)                # row    names
 
@@ -297,16 +297,12 @@ class XPRESS(ConicSolver):
         results_dict = {
 
             'problem':   self.prob_,
-            'status':    self.prob_.getProbStatus(),
-            'obj_value': self.prob_.getObjVal(),
+            'status':    self.prob_.attributes.solstatus,
+            'obj_value': self.prob_.attributes.objval,
         }
 
-        status_map_lp, status_map_mip = get_status_maps()
-
-        if 'mip_' in self.prob_.getProbStatusString():
-            status = status_map_mip[results_dict['status']]
-        else:
-            status = status_map_lp[results_dict['status']]
+        status_map = get_status_map()
+        status = status_map[results_dict['status']]
 
         results_dict[s.XPRESS_TROW] = transf2Orig
 
@@ -315,7 +311,7 @@ class XPRESS(ConicSolver):
         if status in s.SOLUTION_PRESENT:
             results_dict['x'] = self.prob_.getSolution()
             if not (data[s.BOOL_IDX] or data[s.INT_IDX]):
-                results_dict['y'] = - np.array(self.prob_.getDual())
+                results_dict['y'] = - np.array(self.prob_.getDuals())
 
         elif status == s.INFEASIBLE and 'save_iis' in solver_opts and solver_opts['save_iis'] != 0:
 
@@ -361,12 +357,8 @@ class XPRESS(ConicSolver):
         # Generate solution.
         solution = {}
 
-        status_map_lp, status_map_mip = get_status_maps()
-
-        if data[s.BOOL_IDX] or data[s.INT_IDX]:
-            solution[s.STATUS] = status_map_mip[results_dict['status']]
-        else:
-            solution[s.STATUS] = status_map_lp[results_dict['status']]
+        status_map = get_status_map()
+        solution[s.STATUS] = status_map[results_dict['status']]
 
         if solution[s.STATUS] in s.SOLUTION_PRESENT:
 
@@ -380,7 +372,7 @@ class XPRESS(ConicSolver):
         solution[s.XPRESS_IIS] = results_dict[s.XPRESS_IIS]
         solution[s.XPRESS_TROW] = results_dict[s.XPRESS_TROW]
 
-        solution['getObjVal'] = self.prob_.getObjVal()
+        solution['getObjVal'] = self.prob_.attributes.objval
 
         solution[s.SOLVE_TIME] = self.prob_.attributes.time
 
@@ -398,37 +390,17 @@ class XPRESS(ConicSolver):
         """
         return CITATION_DICT["XPRESS"]
 
-def get_status_maps():
-    """Create status maps from Xpress to CVXPY
+def get_status_map():
+    """Create status map from Xpress to CVXPY
     """
 
     import xpress as xp
 
-    # Map of Xpress' LP status to CVXPY status.
-    status_map_lp = {
-
-        xp.lp_unstarted:       s.SOLVER_ERROR,
-        xp.lp_optimal:         s.OPTIMAL,
-        xp.lp_infeas:          s.INFEASIBLE,
-        xp.lp_cutoff:          s.OPTIMAL_INACCURATE,
-        xp.lp_unfinished:      s.OPTIMAL_INACCURATE,
-        xp.lp_unbounded:       s.UNBOUNDED,
-        xp.lp_cutoff_in_dual:  s.OPTIMAL_INACCURATE,
-        xp.lp_unsolved:        s.OPTIMAL_INACCURATE,
-        xp.lp_nonconvex:       s.SOLVER_ERROR
+    status_map = {
+        xp.SolStatus.NOTFOUND:        s.SOLVER_ERROR,
+        xp.SolStatus.OPTIMAL:         s.OPTIMAL,
+        xp.SolStatus.INFEASIBLE:      s.INFEASIBLE,
+        xp.SolStatus.UNBOUNDED:       s.UNBOUNDED,
     }
 
-    # Same map, for MIPs
-    status_map_mip = {
-
-        xp.mip_not_loaded:     s.SOLVER_ERROR,
-        xp.mip_lp_not_optimal: s.SOLVER_ERROR,
-        xp.mip_lp_optimal:     s.SOLVER_ERROR,
-        xp.mip_no_sol_found:   s.SOLVER_ERROR,
-        xp.mip_solution:       s.OPTIMAL_INACCURATE,
-        xp.mip_infeas:         s.INFEASIBLE,
-        xp.mip_optimal:        s.OPTIMAL,
-        xp.mip_unbounded:      s.UNBOUNDED
-    }
-
-    return (status_map_lp, status_map_mip)
+    return status_map
