@@ -74,6 +74,7 @@ GLPK_MI = "GLPK_MI"
 GLOP = "GLOP"
 CBC = "CBC"
 COPT = "COPT"
+CUOPT = "CUOPT"
 ECOS = "ECOS"
 ECOS_BB = "ECOS_BB"
 SCS = "SCS"
@@ -84,6 +85,7 @@ OSQP = "OSQP"
 PIQP = "PIQP"
 PROXQP = "PROXQP"
 QOCO = "QOCO"
+QPALM = "QPALM"
 CPLEX = "CPLEX"
 MOSEK = "MOSEK"
 XPRESS = "XPRESS"
@@ -97,11 +99,13 @@ DAQP = "DAQP"
 HIGHS = "HIGHS"
 MPAX = "MPAX"
 IPOPT = "IPOPT"
+KNITRO = "KNITRO"
+COSMO = "COSMO"
 SOLVERS = [CLARABEL, ECOS, CVXOPT, GLOP, GLPK, GLPK_MI,
            SCS, SDPA, GUROBI, OSQP, CPLEX,
-           MOSEK, CBC, COPT, XPRESS, PIQP, PROXQP, QOCO,
+           MOSEK, CBC, COPT, XPRESS, PIQP, PROXQP, QOCO, QPALM,
            NAG, PDLP, SCIP, SCIPY, DAQP, HIGHS, MPAX,
-           CUCLARABEL, IPOPT]
+           CUCLARABEL, CUOPT, KNITRO, COSMO, IPOPT]
 
 # Xpress-specific items
 XPRESS_IIS = "XPRESS_IIS"
