@@ -16,6 +16,7 @@ limitations under the License.
 
 import math
 import re
+import sys
 import unittest
 
 import numpy as np
@@ -652,8 +653,20 @@ class TestCuClarabel(BaseTest):
     def test_clarabel_pcp_2(self) -> None:
         StandardTestSOCPs.test_socp_2(solver='CUCLARABEL')
 
+def is_mosek_available():
+    """Check if MOSEK is installed and a license is available."""
+    if 'MOSEK' not in INSTALLED_SOLVERS:
+        return False
+    try:
+        import mosek  # type: ignore
+        env = mosek.Env()
+        # Try to get license status (returns 0 if OK)
+        status = env.getlicense()
+        return status == mosek.rescode.ok
+    except Exception:
+        return False
 
-@unittest.skipUnless('MOSEK' in INSTALLED_SOLVERS, 'MOSEK is not installed.')
+@unittest.skipUnless(is_mosek_available(), 'MOSEK is not installed or license is not available.')
 class TestMosek(unittest.TestCase):
 
     def test_mosek_lp_0(self) -> None:
@@ -830,6 +843,7 @@ class TestMosek(unittest.TestCase):
     def test_mosek_sdp_power(self) -> None:
         """Test the problem in issue #2128"""
         StandardTestMixedCPs.test_sdp_pcp_1(solver='MOSEK')
+
 
     def test_power_portfolio(self) -> None:
         """Test the portfolio problem in issue #2042"""
@@ -1086,7 +1100,8 @@ def fflush() -> None:
 
 # We can't inherit from unittest.TestCase since we access some advanced pytest features.
 # As a result, we use the pytest skipif decorator instead of unittest.skipUnless.
-@pytest.mark.skipif('CBC' not in INSTALLED_SOLVERS, reason='CBC is not installed.')
+@pytest.mark.skipif(('CBC' not in INSTALLED_SOLVERS) or sys.platform.startswith("win"),
+                    reason='CBC is not installed or tests are being run on Windows.')
 class TestCBC:
 
     def _cylp_checks_isProvenInfeasible():
@@ -2302,11 +2317,18 @@ class TestHIGHS:
         [StandardTestLPs.test_lp_0, StandardTestLPs.test_mi_lp_0],
     )
     def test_highs_options(self, problem) -> None:
-        with pytest.raises(AttributeError):
+        with pytest.raises(ValueError):
             problem(solver=cp.HIGHS, highs_options={"invalid_highs_option": None})
 
-        with pytest.raises(AttributeError):
+        with pytest.raises(ValueError):
             problem(solver=cp.HIGHS, invalid_highs_option=None)
+
+        invalid_value = "_invalid_value_"
+        with pytest.raises(ValueError):
+            problem(solver=cp.HIGHS, highs_options={"threads": invalid_value})
+
+        with pytest.raises(ValueError):
+            problem(solver=cp.HIGHS, threads=invalid_value)
 
         # Duplicate options
         with pytest.raises(TypeError):
@@ -2322,8 +2344,11 @@ class TestHIGHS:
             "mip_rel_gap": 1.0,
             "primal_feasibility_tolerance": 1e-3,
             "dual_feasibility_tolerance": 1e-3,
+            # advanced option, not available on HighsOptions
+            "presolve_rule_off": 2**10,
         }
         problem(solver=cp.HIGHS, highs_options=highs_options)
+
 
 
 class TestAllSolvers(BaseTest):
@@ -2352,9 +2377,14 @@ class TestAllSolvers(BaseTest):
         prob = cp.Problem(cp.Minimize(cp.norm(self.x, 1) + 1.0), [self.x == 0])
         for solver in SOLVER_MAP_CONIC.keys():
             if solver in INSTALLED_SOLVERS:
-                prob.solve(solver=solver)
-                self.assertAlmostEqual(prob.value, 1.0)
-                self.assertItemsAlmostEqual(self.x.value, [0, 0])
+                if solver is cp.MOSEK and not is_mosek_available():
+                    pass
+                elif solver is cp.KNITRO and not is_knitro_available():
+                    pass
+                else:
+                    prob.solve(solver=solver)
+                    self.assertAlmostEqual(prob.value, 1.0)
+                    self.assertItemsAlmostEqual(self.x.value, [0, 0])
             else:
                 with self.assertRaises(Exception) as cm:
                     prob.solve(solver=solver)
@@ -2362,8 +2392,11 @@ class TestAllSolvers(BaseTest):
 
         for solver in SOLVER_MAP_QP.keys():
             if solver in INSTALLED_SOLVERS:
-                prob.solve(solver=solver)
-                self.assertItemsAlmostEqual(self.x.value, [0, 0])
+                if solver is cp.KNITRO and not is_knitro_available():
+                    pass
+                else:
+                    prob.solve(solver=solver)
+                    self.assertItemsAlmostEqual(self.x.value, [0, 0])
             else:
                 with self.assertRaises(Exception) as cm:
                     prob.solve(solver=solver)
@@ -2377,9 +2410,14 @@ class TestAllSolvers(BaseTest):
             with pytest.raises(cp.error.SolverError, match="You need a mixed-integer "
                                                            "solver for this model"):
                 prob.solve()
-        else:
-            prob.solve()
+        elif is_mosek_available() and cp.MOSEK in INSTALLED_MI_SOLVERS:
+            prob.solve(solver=cp.MOSEK)
             self.assertItemsAlmostEqual(x.value, [0, 0])
+        elif cp.HIGHS in INSTALLED_MI_SOLVERS:
+            prob.solve(solver=cp.HIGHS)
+            self.assertItemsAlmostEqual(x.value, [0, 0])
+        else:
+            pass
 
 
 @unittest.skipUnless('ECOS' in INSTALLED_SOLVERS, 'ECOS_BB is not installed.')
@@ -2391,13 +2429,17 @@ class TestECOS_BB(unittest.TestCase):
         x = cp.Variable(1, name='x', integer=True)
         objective = cp.Minimize(cp.sum(x))
         prob = cp.Problem(objective, [x >= 0])
-        if INSTALLED_MI_SOLVERS != [cp.ECOS_BB]:
+        if INSTALLED_MI_SOLVERS != [cp.ECOS_BB] and is_mosek_available():
             prob.solve()
             assert prob.solver_stats.solver_name != cp.ECOS_BB
+        # The optional solvers build will always have MOSEK installed.
+        """
         else:
-            with pytest.raises(cp.error.SolverError, match="You need a mixed-integer "
-                                                           "solver for this model"):
-                prob.solve()
+            if not is_mosek_available():
+                with pytest.raises(cp.error.SolverError, match="You need a mixed-integer "
+                                                            "solver for this model"):
+                    prob.solve()
+        """
 
     def test_ecos_bb_lp_0(self) -> None:
         StandardTestLPs.test_lp_0(solver='ECOS_BB')
@@ -2620,3 +2662,317 @@ class TestCOPT(unittest.TestCase):
 
         # Valid arg.
         problem.solve(solver=cp.COPT, feastol=1e-9)
+
+
+@unittest.skipUnless('COSMO' in INSTALLED_SOLVERS, 'COSMO is not installed.')
+class TestCOSMO(BaseTest):
+    """Unit tests for COSMO solver interface."""
+    
+    def setUp(self) -> None:
+        self.a = cp.Variable(name='a')
+        self.b = cp.Variable(name='b')
+        self.c = cp.Variable(name='c')
+
+        self.x = cp.Variable(2, name='x')
+        self.y = cp.Variable(3, name='y')
+        self.z = cp.Variable(2, name='z')
+
+        self.A = cp.Variable((2, 2), name='A')
+        self.B = cp.Variable((2, 2), name='B')
+        self.C = cp.Variable((3, 2), name='C')
+
+    def test_cosmo_options(self) -> None:
+        """Test that COSMO solver options work."""
+        # Test basic options with simple problem
+        prob = cp.Problem(cp.Minimize(cp.norm(self.x, 1) + 1.0), [self.x == 0])
+        
+        # Test with default settings
+        prob.solve(solver=cp.COSMO, verbose=True)
+        self.assertAlmostEqual(prob.value, 1.0)
+        self.assertItemsAlmostEqual(self.x.value, [0, 0])
+
+        # Test with custom settings
+        prob.solve(solver=cp.COSMO, eps_abs=1e-4, verbose=True)
+        self.assertAlmostEqual(prob.value, 1.0)
+        self.assertItemsAlmostEqual(self.x.value, [0, 0])
+
+    def test_cosmo_lp_0(self) -> None:
+        StandardTestLPs.test_lp_0(solver='COSMO')
+
+    def test_cosmo_lp_1(self) -> None:
+        StandardTestLPs.test_lp_1(solver='COSMO')
+
+    def test_cosmo_lp_2(self) -> None:
+        StandardTestLPs.test_lp_2(solver='COSMO')
+
+    def test_cosmo_lp_3(self) -> None:
+        StandardTestLPs.test_lp_3(solver='COSMO')
+
+    def test_cosmo_lp_4(self) -> None:
+        StandardTestLPs.test_lp_4(solver='COSMO')
+    
+    def test_cosmo_lp_5(self) -> None:
+        StandardTestLPs.test_lp_5(solver='COSMO')
+
+    def test_cosmo_qp_0(self) -> None:
+        StandardTestQPs.test_qp_0(solver='COSMO')
+
+    def test_cosmo_socp_0(self) -> None:
+        StandardTestSOCPs.test_socp_0(solver='COSMO')
+
+    def test_cosmo_socp_1(self) -> None:
+        StandardTestSOCPs.test_socp_1(solver='COSMO')
+
+    def test_cosmo_socp_2(self) -> None:
+        StandardTestSOCPs.test_socp_2(solver='COSMO')
+
+    def test_cosmo_socp_3(self) -> None:
+        # axis 0
+        StandardTestSOCPs.test_socp_3ax0(solver='COSMO')
+        # axis 1 
+        StandardTestSOCPs.test_socp_3ax1(solver='COSMO')
+
+    def test_cosmo_expcone_1(self) -> None:
+        StandardTestECPs.test_expcone_1(solver='COSMO')
+
+    def test_cosmo_exp_soc_1(self) -> None:
+        StandardTestMixedCPs.test_exp_soc_1(solver='COSMO')
+
+    def test_cosmo_pcp_1(self) -> None:
+        StandardTestPCPs.test_pcp_1(solver='COSMO')
+
+    def test_cosmo_pcp_2(self) -> None:
+        StandardTestPCPs.test_pcp_2(solver='COSMO')
+
+    def test_cosmo_pcp_3(self) -> None:
+        StandardTestPCPs.test_pcp_3(solver='COSMO')
+
+    def test_cosmo_sdp_1min(self) -> None:
+        StandardTestSDPs.test_sdp_1min(solver='COSMO')
+    
+    def test_cosmo_sdp_2(self) -> None:
+        places = 3
+        sth = sths.sdp_2()
+        sth.solve(solver='COSMO')
+        sth.verify_objective(places)
+        sth.check_primal_feasibility(places)
+        sth.verify_primal_values(places)
+        sth.check_complementarity(places)
+        sth.check_dual_domains(places)
+
+    def test_warm_start(self) -> None:
+        """Test warm starting.
+        """
+        x = cp.Variable(10)
+        obj = cp.Minimize(cp.sum(cp.exp(x)))
+        prob = cp.Problem(obj, [cp.sum(x) == 1])
+        result = prob.solve(solver=cp.COSMO)
+        time = prob.solver_stats.solve_time
+        result2 = prob.solve(solver=cp.COSMO, warm_start=True)
+        time2 = prob.solver_stats.solve_time
+        self.assertAlmostEqual(result2, result, places=2)
+        print(time > time2)
+
+def is_knitro_available():
+    """Check if KNITRO is installed and a license is available."""
+    if 'KNITRO' not in INSTALLED_SOLVERS:
+        return False
+    try:
+        import knitro  # type: ignore
+        # Try to create and delete a Knitro solver instance
+        kc = knitro.KN_new()
+        if kc is None:
+            return False
+        knitro.KN_free(kc)
+        return True
+    except Exception:
+        return False
+
+@unittest.skipUnless(is_knitro_available(), 'KNITRO is not installed or license is not available.')
+class TestKNITRO(BaseTest):
+
+    def test_knitro_lp_0(self) -> None:
+        StandardTestLPs.test_lp_0(solver=cp.KNITRO)
+
+    def test_knitro_lp_1(self) -> None:
+        StandardTestLPs.test_lp_1(solver=cp.KNITRO)
+
+    def test_knitro_lp_2(self) -> None:
+        StandardTestLPs.test_lp_2(solver=cp.KNITRO)
+
+    def test_knitro_lp_3(self) -> None:
+        StandardTestLPs.test_lp_3(solver=cp.KNITRO)
+
+    def test_knitro_lp_4(self) -> None:
+        StandardTestLPs.test_lp_4(solver=cp.KNITRO)
+
+    def test_knitro_lp_5(self) -> None:
+        StandardTestLPs.test_lp_5(solver=cp.KNITRO)
+
+    def test_knitro_lp_6(self) -> None:
+        StandardTestLPs.test_lp_6(solver=cp.KNITRO)
+
+    def test_knitro_lp_bound_attr(self) -> None:
+        StandardTestLPs.test_lp_bound_attr(solver=cp.KNITRO)
+
+    def test_knitro_socp_0(self) -> None:
+        StandardTestSOCPs.test_socp_0(solver=cp.KNITRO)
+
+    def test_knitro_socp_1(self) -> None:
+        StandardTestSOCPs.test_socp_1(solver=cp.KNITRO)
+
+    def test_knitro_socp_2(self) -> None:
+        StandardTestSOCPs.test_socp_2(solver=cp.KNITRO)
+
+    def test_knitro_socp_3(self) -> None:
+        # axis 0
+        StandardTestSOCPs.test_socp_3ax0(solver=cp.KNITRO)
+        # axis 1
+        StandardTestSOCPs.test_socp_3ax1(solver=cp.KNITRO)
+
+    def test_knitro_socp_bounds_attr(self) -> None:
+        StandardTestSOCPs.test_socp_bounds_attr(solver=cp.KNITRO)
+
+    def test_knitro_mi_lp_0(self) -> None:
+        StandardTestLPs.test_mi_lp_0(solver=cp.KNITRO)
+
+    def test_knitro_mi_lp_1(self) -> None:
+        StandardTestLPs.test_mi_lp_1(solver=cp.KNITRO)
+
+    def test_knitro_mi_lp_2(self) -> None:
+        StandardTestLPs.test_mi_lp_2(solver=cp.KNITRO)
+
+    def test_knitro_mi_lp_3(self) -> None:
+        StandardTestLPs.test_mi_lp_3(solver=cp.KNITRO)
+
+    def test_knitro_mi_lp_5(self) -> None:
+        StandardTestLPs.test_mi_lp_5(solver=cp.KNITRO)
+
+    def test_knitro_mi_socp_1(self) -> None:
+        StandardTestSOCPs.test_mi_socp_1(solver=cp.KNITRO)
+
+    def test_knitro_mi_socp_2(self) -> None:
+        StandardTestSOCPs.test_mi_socp_2(solver=cp.KNITRO)
+
+    def test_knitro_qp_0(self) -> None:
+        StandardTestQPs.test_qp_0(solver=cp.KNITRO)
+
+    def test_knitro_expcone_1(self) -> None:
+        StandardTestECPs.test_expcone_1(solver=cp.KNITRO)
+
+    def test_knitro_pcp_1(self) -> None:
+        StandardTestPCPs.test_pcp_1(solver=cp.KNITRO)
+
+    def test_knitro_pcp_2(self) -> None:
+        StandardTestPCPs.test_pcp_2(solver=cp.KNITRO)
+
+    def test_knitro_pcp_3(self) -> None:
+        StandardTestPCPs.test_pcp_3(solver=cp.KNITRO)
+
+    def test_knitro_mi_pcp_0(self) -> None:
+        StandardTestPCPs.test_mi_pcp_0(solver=cp.KNITRO)
+
+    def test_knitro_sdp_1min(self) -> None:
+        StandardTestSDPs.test_sdp_1min(solver=cp.KNITRO)
+
+    def test_knitro_sdp_1max(self) -> None:
+        StandardTestSDPs.test_sdp_1max(solver=cp.KNITRO)
+
+    def test_knitro_sdp_2(self) -> None:
+        StandardTestSDPs.test_sdp_2(solver=cp.KNITRO)
+
+    def test_knitro_exp_soc_1(self) -> None:
+        StandardTestMixedCPs.test_exp_soc_1(solver=cp.KNITRO)
+
+    def test_knitro_sdp_pcp_1(self) -> None:
+        StandardTestMixedCPs.test_sdp_pcp_1(solver=cp.KNITRO)
+
+    def test_knitro_params(self) -> None:
+        n = 10
+        m = 4
+        np.random.seed(0)
+        A = np.random.randn(m, n)
+        x = np.random.randn(n)
+        y = A.dot(x)
+
+        # Solve a simple basis pursuit problem for testing purposes.
+        z = cp.Variable(n)
+        objective = cp.Minimize(cp.norm1(z))
+        constraints = [A @ z == y]
+        problem = cp.Problem(objective, constraints)
+
+        with self.assertRaises(Exception):
+            opts = {"a": "invalid"}
+            problem.solve(solver=cp.KNITRO, **opts)
+
+        with self.assertRaises(Exception):
+            opts = {"algorithm": "invalid"}
+            problem.solve(solver=cp.KNITRO, **opts)
+
+        opts = {"algorithm": 0}
+        problem.solve(solver=cp.KNITRO, **opts)
+
+@unittest.skipUnless("CUOPT" in INSTALLED_SOLVERS, "CUOPT is not installed.")
+class TestCUOPT(unittest.TestCase):
+
+    import os
+    kwargs={"pdlp_solver_mode": os.environ.get("CUOPT_PDLP_SOLVER_MODE", "Stable2"),
+            "solver_method": os.environ.get("CUOPT_SOLVER_METHOD", 0)
+            }
+
+    def test_cuopt_lp_0(self) -> None:
+        StandardTestLPs.test_lp_0(solver="CUOPT", duals=True, places=4, **TestCUOPT.kwargs)
+
+    def test_cuopt_lp_1(self) -> None:
+        StandardTestLPs.test_lp_1(solver="CUOPT", duals=True, places=4, **TestCUOPT.kwargs)
+
+    def test_cuopt_lp_2(self) -> None:
+        StandardTestLPs.test_lp_2(solver="CUOPT", duals=True, places=4, **TestCUOPT.kwargs)
+
+    def test_cuopt_lp_3(self) -> None:
+        StandardTestLPs.test_lp_3(solver="CUOPT", duals=True, places=4, **TestCUOPT.kwargs)
+
+    def test_cuopt_lp_4(self) -> None:
+        # In this case cuopt throws an exception because there are crossing
+        # variable bounds x <= 0 and x >= 1
+        try:
+            StandardTestLPs.test_lp_4(solver="CUOPT", duals=True, places=4, **TestCUOPT.kwargs)
+        except Exception as e:
+            assert "crossing bounds" in str(e)
+
+    def test_cuopt_lp_5(self) -> None:
+        StandardTestLPs.test_lp_5(solver='CUOPT', duals=True, places=4, **TestCUOPT.kwargs)
+
+    def test_cuopt_lp_6(self) -> None:
+        StandardTestLPs.test_lp_5(solver='CUOPT', duals=True, places=4, **TestCUOPT.kwargs)
+
+    def test_cuopt_lp_7(self) -> None:
+        StandardTestLPs.test_lp_5(solver='CUOPT', duals=True, places=4, **TestCUOPT.kwargs)
+
+    def test_cuopt_mi_lp_0(self) -> None:
+        StandardTestLPs.test_mi_lp_0(solver='CUOPT', **TestCUOPT.kwargs)
+
+    def test_cuopt_mi_lp_1(self) -> None:
+        StandardTestLPs.test_mi_lp_1(solver='CUOPT', **TestCUOPT.kwargs)
+
+    def test_cuopt_mi_lp_2(self) -> None:
+        StandardTestLPs.test_mi_lp_2(solver='CUOPT', **TestCUOPT.kwargs)
+
+    def test_cuopt_mi_lp_3(self) -> None:
+        TestCUOPT.kwargs["time_limit"] = 5
+        StandardTestLPs.test_mi_lp_3(solver='CUOPT', **TestCUOPT.kwargs)
+        del TestCUOPT.kwargs["time_limit"]
+
+    # This is an unconstrained problem, which cuopt doesn't handle.
+    # Error message from cvxpy should be returned
+    def test_cuopt_mi_lp_4(self) -> None:
+        try:
+            StandardTestLPs.test_mi_lp_4(solver='CUOPT', **TestCUOPT.kwargs)
+        except Exception as e:
+            assert "there are not enough constraints in the problem" in str(e)
+
+    def test_cuopt_mi_lp_5(self) -> None:
+        StandardTestLPs.test_mi_lp_5(solver='CUOPT', **TestCUOPT.kwargs, time_limit=5)
+
+    def test_cuopt_mi_lp_7(self) -> None:
+        StandardTestLPs.test_mi_lp_5(solver='CUOPT', **TestCUOPT.kwargs, time_limit=5)
