@@ -1433,13 +1433,10 @@ class TestProblem(BaseTest):
     def test_solver_error_raised_on_failure(self) -> None:
         """Tests that a SolverError is raised when a solver fails.
         """
-        A = numpy.random.randn(40, 40)
-        b = cp.matmul(A, numpy.random.randn(40))
-
         with self.assertRaises(SolverError):
-            Problem(cp.Minimize(
-                cp.sum_squares(cp.matmul(A, cp.Variable(40)) - b))).solve(
-                solver=s.OSQP, max_iter=1)
+            Problem(cp.Minimize(cp.quad_form(cp.Variable(1) + 1, np.array([[-1]]), True))).solve(
+                solver=s.OSQP
+            )
 
     def test_solve_solver_path(self) -> None:
         """
@@ -1468,9 +1465,9 @@ class TestProblem(BaseTest):
         solvers = [(s.OSQP, {'max_iter':1})]
 
         with self.assertRaises(SolverError):
-            Problem(cp.Minimize(
-                cp.sum_squares(cp.matmul(A, cp.Variable(40)) - b))).solve(
-                solver_path=solvers)
+            Problem(cp.Minimize(cp.quad_form(cp.Variable(1) + 1, np.array([[-1]]), True))).solve(
+                solver_path=solvers
+            )
 
         # invalid input, raise ValueError
         solvers_invalid_inner_input = [{'str':{}}, 'str', [], [1], [()], [(1)], [(1,{})],
